# swagger.yaml
openapi: 3.0.1
info:
  title: Get Science API
  description: API платформы для научных и образовательных мероприятий
  version: 1.0.0
  contact:
    name: Get Science Team
    url: https://get-science-events.ru
servers:
  - url: https://get-science-events.ru/api/v1
    description: Основной сервер
...
paths:
  /applications/submit:
    post:
      tags:
        - applications
      summary: Создание новой заявки на участие в мероприятии
      description: Создает новую заявку на участие в мероприятии
      operationId: submitApplication
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                applicationRequest:
                  $ref: '#/components/schemas/ApplicationRequest'
                fileApplicationFileMetadataDTO:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApplicationFileMetadataDTO'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Успешное создание заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Ошибка валидации данных
  /applications/{applicationId}:
    put:
      tags:
        - applications
      summary: Обновление существующей заявки
      description: Обновляет существующую заявку
      operationId: updateApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                applicationRequest:
                  $ref: '#/components/schemas/ApplicationRequest'
                fileApplicationFileMetadataDTO:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApplicationFileMetadataDTO'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Успешное обновление заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Заявка не найдена
    get:
      tags:
        - applications
      summary: Получение информации о заявке
      description: Получает информацию о заявке по её идентификатору
      operationId: getApplicationById
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о заявке успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Заявка не найдена
    delete:
      tags:
        - applications
      summary: Удаление заявки
      description: Удаляет заявку по её идентификатору
      operationId: deleteApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Заявка успешно удалена
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Заявка не найдена
  /applications/{applicationId}/update-organizer:
    put:
      tags:
        - applications
      summary: Обновление заявки организатором
      description: Позволяет организатору мероприятия обновить заявку
      operationId: updateApplicationByOrganizer
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRequest'
      responses:
        '200':
          description: Заявка успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Заявка не найдена
  /applications/{applicationId}/with-applicant:
    get:
      tags:
        - applications
      summary: Получение детальной информации о заявке с данными заявителя
      description: Получает детальную информацию о заявке вместе с данными заявителя
      operationId: getApplicationWithApplicant
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailWithApplicantResponse'
        '404':
          description: Заявка не найдена
  /applications/{applicationId}/with-organizer:
    get:
      tags:
        - applications
      summary: Получение детальной информации о заявке с данными организатора
      description: Получает детальную информацию о заявке вместе с данными организатора
      operationId: getApplicationWithOrganizer
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailWithOrganizerResponse'
        '404':
          description: Заявка не найдена
  /applications/event/{eventId}:
    get:
      tags:
        - applications
      summary: Получение списка заявок для мероприятия
      description: Получает список всех заявок для указанного мероприятия
      operationId: getApplicationsByEvent
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Мероприятие не найдено
  /applications/profile/{profileId}:
    get:
      tags:
        - applications
      summary: Получение списка заявок для профиля
      description: Получает список всех заявок для указанного профиля
      operationId: getApplicationsByProfile
      parameters:
        - name: profileId
          in: path
          description: Идентификатор профиля
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Профиль не найден
  /applications/organizer:
    get:
      tags:
        - applications
      summary: Получение списка заявок для текущего организатора
      description: Получает список всех заявок для текущего организатора мероприятий
      operationId: getApplicationsByOrganizer
      responses:
        '200':
          description: Список заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '401':
          description: Не авторизован
  /applications/applicant:
    get:
      tags:
        - applications
      summary: Получение списка заявок для текущего заявителя
      description: Получает список всех заявок для текущего заявителя
      operationId: getApplicationsByApplicant
      responses:
        '200':
          description: Список заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '401':
          description: Не авторизован
...
  /files/applications/{applicationId}:
    get:
      tags:
        - files
      summary: Получение файлов заявки
      description: Получает список файлов, прикрепленных к заявке
      operationId: getFileApplicationsByApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список файлов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileApplicationResponse'
        '404':
          description: Заявка не найдена
...
  /files/applications/{fileId}:
    delete:
      tags:
        - files
      summary: Удаление файла заявки
      description: Удаляет файл, прикрепленный к заявке
      operationId: deleteFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл успешно удален
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Файл не найден
...
components:
  schemas:
    ApplicationRequest:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
          nullable: true
        status:
          type: string
          description: Статус заявки
        submissionDate:
          type: string
          format: date-time
          description: Дата подачи заявки
        message:
          type: string
          description: Сообщение от заявителя
          nullable: true
        isObserver:
          type: boolean
          description: Является ли наблюдателем
        fileApplications:
          type: array
          description: Файлы заявки
          items:
            $ref: '#/components/schemas/FileApplicationRequest'
        verdict:
          type: string
          description: Решение по заявке
          nullable: true
    ApplicationResponse:
      type: object
      properties:
        applicationId:
          type: integer
          format: int64
          description: Идентификатор заявки
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        eventName:
          type: string
          description: Название мероприятия
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
        status:
          type: string
          description: Статус заявки
        submissionDate:
          type: string
          format: date-time
          description: Дата подачи заявки
        message:
          type: string
          description: Сообщение от заявителя
          nullable: true
        isObserver:
          type: boolean
          description: Является ли наблюдателем
        verdict:
          type: string
          description: Решение по заявке
          nullable: true
        fileApplications:
          type: array
          description: Файлы заявки
          items:
            $ref: '#/components/schemas/FileApplicationResponse'
    ApplicationFileMetadataDTO:
      type: object
      properties:
        type:
          type: string
          description: Тип файла
        isEncrypted:
          type: boolean
          description: Зашифрован ли файл
        docRequiredId:
          type: integer
          format: int64
          description: Идентификатор требуемого документа
...
    FileApplicationRequest:
      type: object
      properties:
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        isEncryptionEnabled:
          type: boolean
          description: Включено ли шифрование
    FileApplicationResponse:
      type: object
      properties:
        fileId:
          type: integer
          format: int64
          description: Идентификатор файла
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        docRequired:
          $ref: '#/components/schemas/DocRequiredResponse'
          nullable: true
          description: Требуемый документ
        isEncryptionEnabled:
          type: boolean
          description: Включено ли шифрование
...
    DocRequiredResponse:
      type: object
      properties:
        docRequiredId:
          type: integer
          format: int64
          description: Идентификатор требуемого документа
        name:
          type: string
          description: Название требуемого документа
        description:
          type: string
          description: Описание требуемого документа
          nullable: true
        required:
          type: boolean
          description: Обязателен ли документ
...
    ProfileResponse:
      type: object
      properties:
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
        firstName:
          type: string
          description: Имя пользователя
        lastName:
          type: string
          description: Фамилия пользователя
        aboutMe:
          type: string
          description: Информация о пользователе
          nullable: true
        avatarUrl:
          type: string
          description: URL аватара
          nullable: true
        avatarPresignedUrl:
          type: string
          description: Предварительно подписанный URL аватара
          nullable: true
        role:
          type: string
          description: Роль пользователя
        email:
          type: string
          format: email
          description: Email пользователя
        reviewEvents:
          type: array
          description: Мероприятия, для которых пользователь является рецензентом
          items:
            $ref: '#/components/schemas/EventShortDto'
        coownerEvents:
          type: array
          description: Мероприятия, для которых пользователь является соорганизатором
          items:
            $ref: '#/components/schemas/EventShortDto'
        isActive:
          type: boolean
          description: Активен ли профиль
          nullable: true
    EventShortDto:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        title:
          type: string
          description: Название мероприятия
...
    ApplicationDetailWithApplicantResponse:
      type: object
      properties:
        application:
          $ref: '#/components/schemas/ApplicationResponse'
        applicant:
          $ref: '#/components/schemas/ProfileResponse'
    ApplicationDetailWithOrganizerResponse:
      type: object
      properties:
        application:
          $ref: '#/components/schemas/ApplicationResponse'
        organizer:
          $ref: '#/components/schemas/ProfileResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []